Use latest yottadb-base docker image and install go into it
Change from using the public go docker image to using yottadb-base:latest
so yottadb is installed by default - removing a large compilation task
from the build process. This has the effect that we now need to install
go at every build stage, but that is far faster since we are using binaries
to do that step.

Also, fix link for code coverage report to go to the jobs page instead
of gitlab pages due to gitlab pages being unsupported for subgroups