Remove unit test with timing constraints from CI, add YDB_OK to errors
TestMiscGoTimers has cases where the MUMPS hang takes upwards of
1.2-1.5s to return, versus the expected 1s. Overnight testing on an
internal test runner, with no load other than concurrent runs of this,
show the failure at a rate of ~5/328 (1.5%). Given that we have only
seen this in the CI docker container, we are removing it from the CI
runs. It will still be run in the YDB test framework.