Fix SetValStrLit/SetValStr/SetValBAry methods of BufferT object to leave len_used untouched in case of INVSTRLEN; Fix Dump() method for BufferT and BufferTArray objects to not go over allocated buffer; Add tests for these scenarios (used to SIG-11)
Keeping len_used untouched in case of an INVSTRLEN error in the BufferT object methods keeps
the code consistent with the documentation. Also reduces the chances the BufferT object
gets into an out-of-design state (len_used > len_alloc).

Note that len_used is greater than len_alloc is not possible in Go Wrapper (because the only
place where it was possible BufferT object is now fixed as part of this commit) but is
possible in C SimpleAPI so needs to be handled by Dump() methods of BufferT and BufferTArray objects.