Do not create a new database if <go test> is being invoked inside the test system
This is because test system randomly sets various parameters while creating the database
and using the test system created database (using dbcreate.csh) is better for more test
coverage. If "go test" is invoked outside the test system, we do want to create a new
database to avoid polluting globals in a pre-existing/current database of the invoking user.