[#8] Change NodeNextST()/NodePrevST() processing so output subscript array can be the array of the input key
This is not necessarily a good programming tactic because if an error occurs like INVSTRLEN, the input key may be
completely or partially destroyed but be that as it may be, it is more efficient. The task is accomplished by
buffering the input value for the size of the output array (number of output subscripts possible) and when this
value is modified by the Simple API call, store that value back into the key's subscript array.

Change also removed an unneeded code line in both NodeNextE()/NodePrevE() that was setting a value to itself.

A consistency change was made to change all references to "elems*" (case irrelevant) to "elem*" to make variable
names similar to routine names.

Also made a change to gen_error_codes.sh to allow the YDB_MAX_ERRORMSG value to be picked up out of libyottadb.h
and made available in the generated error_code.go file. This was needed for the next item.

Lastly, the largest part of this change is removing the 'C.' prefix on all the error messages and other constants
and relying on the values in error_codes.go instead. This actually shrunk the built executable a few percent.