[#18] Use calloc() instead of malloc() for C allocations inside Go
https://golang.org/cmd/cgo/#hdr-Passing_pointers has the following to say about this issue.

Note: the current implementation has a bug. While Go code is permitted to write nil or a C pointer
(but not a Go pointer) to C memory, the current implementation may sometimes cause a runtime error
if the contents of the C memory appear to be a Go pointer. Therefore, avoid passing uninitialized
C memory to Go code if the Go code is going to store pointer values in it. Zero out the memory in
C before passing it to Go.

The C.calloc() call initializes memory whereas C.malloc() does not initialize memory.
Running the go/randomWalkSimple subtest with this and GOGC env var set to 1 results in no errors
of the following two types whereas previously one would get a failure rate of around 20%.

fatal error: found bad pointer in Go heap (incorrect use of unsafe or cgo?)
fatal error: sweep increased allocation count