[#18] Changes to add runtime.KeepAlive() calls to methods and functions to keep our C allocated storage from havings its anchor blocks garbage collected while the anchored C storage is still in use.
The issue with Finalizers is that they can be driven unexpectedly. The moment the code
is informed that a (Go) variable is no longer needed (which can be pretty much ANYWHERE),
it becomes eligible for garbage collection. What we've seen happen is a Go line pulls the
address of the C storage out of the Go anchor but that reference was the last time the
Go anchor was used so the next line when the C routine is driven is when the garbage
collector is driven that then frees the C storage before it even makes it to the C call
that is going to use it.

These KeepAlive() calls, added after the last need for the C storage, are references to
the anchor block(s) that prevent the garbage collector from identifying our anchor
blocks as GC-able until after the need for them has ended.

The KeepAlive() calls are generally added to routines where parameter or receiver values
or even values created by the routine have pointers to C structures extracted from them.
In some cases the pointers are used to fetch, add, or modify data and in others, the
values are passed to C code so need to be protected. If a value (like an errstr) is
simply being passed to another Go routine, there is no need to protect it so KeepAlive()
calls for that structure are not added in that case.

This commit also removes the "dev" subdirectory which held a tool that was required
by an earlier version of the wrapper but is no longer needed.