[#17] Add error message value substitution for INVSTRLEN errors created by the wrapper
Substitution parameters were added to the INVSTRLEN message since the wrapper was first
created. When this error is created by the YDB C code base, those parameters are properly
substituted but when they are raised by the wrapper itself, the parameters are NOT
substituted generating a fairly confusing message for users.

This support uses the errorFormat() routine added by this support to perform the proper
parameter substitutions for those INVSTRLEN errors generated by the wrapper. It can be
used for other error substitutions in the future.

Modified existing tests that verify INVSTRLEN (from non-SimpleAPI type calls) to also
check that the message is formatted properly verifying the support added here.