[YottaDB/Lang/YDBGo#21] Additional Go wrapper cleanups/fixes
This change has a number of cleanups across several modules. These are mostly in the
form of removing unnecessary casts and pointer dereferences.

One somewhat larger change modifies the unexported variadicPlist struct methods to
anchor a C struct containing an array of void* pointers instead of uintptr_t integers.
Using uintptr types hides pointers from Go which ends up complicating the code
requiring extra unnecessary casts and is not "good Go" but using pointers instead
means we need a method of converting an int/uint64 to a pointer which Go does not
allow one to do so small internal C stub routines are added to do this "conversion".

The KeepAlive changes added in previous commit(s) generally take care of the issues
that the above variadicPlist changes took care of but since the change is still to
have better Go code, the change is being implemented.

There is also a potential fix in ElemLenUsed() in buffer_t_array.go where the return
statement was referencing a pointer to C memory whose anchor was protected by the
KeepAlive() on the preceding line. This was changed so the KeepAlive() is AFTER the
reference (separated from the return statement) to keep the protection intact for
that reference.

Note this change requires an accompanying change in YDB as the variadic plist call entry point
(ydb_call_variadic_plist_func()) has its argument type changed from uintptr_t to void*.