[YottaDB/Lang/YDBGo#6] DataST() and DataE() should not return 0 on an error
This change applies not only to the Go wrapper but also to the Simple API calls ydb_data_s()
and ydb_data_st() (See [YottaDB/DB/YDB#472]). We define the expected return codes now in
libyottadb.h as YDB_DATA_*. One of those codes is YDB_DATA_ERROR which is returned in
*ret_value when ydb_data_s() or ydb_data_st() get an error and return a non-YDB_OK return
code. This value is then returned as the return value of DataST()/DataE().

We do not really recommend that users rely on YDB_DATA_ERROR because the recommended coding
method is to check for error values in where they are returned before trying to look
at the return value. But for those who wish it, we are now providing a code that is not a
valid data-describing return value.

Previously these functions were returning 0 in this error condition which
itself is a valid return value.

This update also includes changes to the error_code.go generation scheme so the new
values also appear in error_codes.go. Also includes are updates to the unit tests to verify
the use of the new codes.