[YottaDB/Lang/YDBGo#21] Revert part of 6edc288016781f03f26d130c1d39d1b46d953097 that changed variadicPlist array from uinptr to void *
An issue on AARCH64 has been identified where we are putting non-pointer values into the variadic
plist void* array. This involves forcing non-pointer values into pointers even when they aren't.
This causes issues in cgo. So revert to the array being uintptr (uintptr_t on the C side) instead.
The symptom seen was a panic with the error that a memory barrier had an invalid pointer.