[#28] Fix one error in a Go test, an incorrect error about too many arguments on 32 bit ARM, and add an extra sanity test - update errors
The timing/go_unit_test subtest that ran the timed tests in the go unit tests started showing an error after YDBGo#28. The
error was about an invalid output parameter. The problem here was the call to CallMT() for routine "TestMGoTimers" had NO
parameters. The way you show no subscripts on an EASY API call is to pass []string{} but we do not use that syntax when
calling M routines. So this was passing a parameter that was not defined in the table and getting the wrong error. An
extra sanity check is verifying that the parameter is defined and if not, it panics.

There was also an issue on ARM 32 bit machine which have an extra parameter since the tptoken (a uint64 value) must be split
across two parms. In calls that take the parameter index, we check for <= the parameter max but in this case, where we are
running SetUsed(), we are using an actual count so the check should be < and not <=.

Also updated error codes from recent commits to YDB.