Fix occasional failures in timing/go_unit_tests subtest (in YDBTest repo)
Occasionally we see the `timing/go_unit_tests` subtest fail during in-house testing
on an x86_64 box with the following diff

```diff
4c4,13
< PASS
---
> --- FAIL: TestMiscGoTimers (2.22s)
>     helpers.go:243: Running a timed test which may fail on a loaded system
>     <autogenerated>:1:
>               Error Trace:    misc_test.go:56
>                                                       asm_amd64.s:1333
>               Error:          Relative error is too high: 0.2 (expected)
>                                       < 0.22028919200000008 (actual)
>               Test:           TestMiscGoTimers
> FAIL
> TEST-E-FAILED : go test lang.yottadb.com/go/yottadb returned failure status of 0
```

The allowed error percentage is `20%` but the actual error seen is `22%`. We have seen other such
failures where the actual error goes as high as `23%`. In all cases, the system was loaded so no
actual issue is suspected.

Given this, we are bumping the allowed error percentage from `20%` to `30%` as part of this commit.
Took this opportunity to also avoid code duplication of `.2` and instead use a variable `allowedErrorPct`
that is initialized to `0.30`.