[Dev version] Update development version to 1.1.1 and update the Go versions tested in pipeline
The Go development versions will now have an "odd" 3rd digit that is one more than the preceding
released (master) version of the wrapper. In these development versions, we will also specify the
minimum YottaDB release as the current master version though this value may or may not revert to
the previous release value if no dependent changes in YottaDB were made.