[YottaDB/DB/YDB#836] Change parameter in ydb_call_variadic_plist() from an intptr_t to gparam_list*
This change was made for the Python wrapper to make it easier for it. This change in parameter type though
caused a ripple effect in the other wrappers so Go and Rust wrappers also need to change so the parameters
they pass match the new type.

Also updated the pipeline script to include additional Go versions to test on.