[#34] Change errors STRUCTNOTALLOCD -> STRUCTUNALLOCD and INVLNPAIRLIST -> INVLKNMPAIRLIST
This change was needed as the two errors originally lived in ydberrors.msg but they were marked
as Go-wrapper-only errors. As part of #34, we have added an internal wrapper error facility for
Go-wrapper-only errors so because the wrapper pulls in the include file giving it access to ALL
YDB errors, we had to change the names of the copies of these errors we put in our local facility
to avoid a name clash with the error list from YDB that the wrapper imports in error_codes.go.

We will remove the STRUCTNOTALLOCD and INVLNPAIRLIST errors from ydberrors.msg in the next release
of YottaDB (r1.36).