[CI] Fix CI issue with go-lint requiring Go 1.18 and subsequent issues with go-lint
A recent change in a 3rd party file pulled in by 'golint' forced golint to be
run with Go 1.18. So for golint only, we override the Go version to 1.18.9 (latest
point release for 1.18). This then brought on another twist in that the 'golint'
executable was no longer found in ${GOPATH}/bin. This was because of the Go Module
support (enforced by default in 1.18) was building the executable elsewhere with
an indeterminable path like /go/pkg/mod/golang.org/x/lint@v0.0.0-20210508222113-6edffad5e616/golint.
The solution was to set GO111MODULE=off in the pipeline, which is how we run locally
until YDBTest#387 is complete.