[CI] Reduce CI testing of Go versions as it is getting a bit burdensome
To this point, we've been testing the first and last Go point release in every
major Go version. This change removes the testing of the first point release in
every major Go version retaining only the last/current point release in every
major Go version. We also changed the version the test is run with from 1.13.0
to the last v1.13 point release v1.13.8.

This change to the GOLANG testing version needs to be reflected in a few other
files besides those in this project. A separate MR updates those files in YDBOcto
repository.