[#51] Bring README up to date for 2024
The README has not been updated for Go Module changes; the Go Ecosystem
moved on on how development is actually done. Modules are the default
development methodology as of Go 1.13.

See https://go.dev/blog/using-go-modules.

YottaDB has had documentation to use environment variable `GO111MODULE`
to use the old behavior, but that's slated to be removed in Go 1.22
(which is the currently released version now).

With this in mind, the main changes in the README are to:

- Introduce the user to current industry Go-standard methods for using
  YottaDB Go Wrapper
- Show new developers how to switch the client Go program from using the
  Go package version to the local version on the file system
- Remove references to `threenp1` application, which doesn't build
  anymore, and even after I spent the time to fix all the problems in
  it, it displays a demo that seems quite different from the
  documentation here.
- Fix Coverage SVG display. It seems that a Gitlab upgrade broke our
  display in the past.
- Add vim to the Docker image.