Simplified SignalNotify API, making it the same as Notify.Signal()
Removed ackChan and notifyWhen for a simpler mechanism. The user now has to
call YDBNotify(signal) from their signal handler rather than sending ack to ackChan and specifying notifyWhen.
(The user is no more likely to forget to call YDBNotify than to forget to send to ackChan.)
Removing ackChan and notifyWhen simplifies the API and will allow removal of our superfluous channels
and watcher Goroutines in a subsequent commit.

Also automatically populate ydbSignalMap for less code repetition and so that YDBSignals can be made public
for documentation purposes.

Change to non-blocking writes to goroutine shutdown channel sigInfo.shutdownNow to be in line with Signal.Notify() behaviour.

Tidu-ups:
- Merge sigNotificationMap into ydbSignalMap as two separate maps were unnecessary.
- Remove functions that are now obsolete: waitForSignalAckWTimeout(), StringSelect().
- Remove notifyWen defines that are now obsolete in yottadb.go so that it is now fairly minimal.
- Fix up several golint complaints.