Fix bug where transaction panic could traverse CGo boundary - bug also present in YDBGo v1
The transaction callback function could previously panic which would unwind the Go stack as well as the
C stack that called it and the Go stack that called that. Traversing the CGo boundary like this can destabilize
the entire process. But sometimes it would work fine: making it a particularly onerous bug.
This has been fixed by recovering any panics in tpCallbackWrapper, returning any errors through
CGo back to Go and then re-panicing the error in Conn.Transaction().

Also added:
- Ability to specify the action to be taken on transaction timeout: commit, rollback, or rollback and issue TPTIMEOUT panic
- Prevented infinite recursion if trying to recover an error message on an error like TPTIMEOUT that has no message.